+++ lib/src/auth0_utilities.dart
@@ -1,5 +1,0 @@
-import 'dart:io' show Platform;
-
-String callbackUri({String bundleId, String domain}) {
-  return '${bundleId.toLowerCase()}://$domain/${Platform.isIOS ? 'ios' : 'android'}/$bundleId/callback';
-}