+++ lib/src/networking/telemetry.dart
@@ -1,0 +1,6 @@
+part of auth0_auth;
+
+const telemetry= {
+  'name': 'flutter_auth0',
+  'version': '1.0.0'
+};